tests/wpt/web-platform-tests/picture-in-picture/request-picture-in-picture-twice.html

<!DOCTYPE html>
<title>Test request Picture-in-Picture on two videos</title>
<meta name="timeout" content="long">
<script src="/common/media.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/picture-in-picture-helpers.js"></script>
<body></body>
<script>
promise_test(async t => {
  const video1 = await loadVideo();
  const video2 = await loadVideo();
  await test_driver.bless('request Picture-in-Picture');
  const promise = video1.requestPictureInPicture();
  await promise_rejects(t, 'NotAllowedError', video2.requestPictureInPicture());
  return promise;
}, 'request Picture-in-Picture consumes user gesture');

promise_test(async t => {
  const video1 = await loadVideo();
  const video2 = await loadVideo();
  await test_driver.bless('request Picture-in-Picture');
  await video1.requestPictureInPicture();
  assert_equals(document.pictureInPictureElement, video1);
  return video2.requestPictureInPicture();
}, 'request Picture-in-Picture does not require user gesture if document.pictureInPictureElement is set');
</script>
